+++ rocketgram/__init__.py
@@ -14,3 +14,6 @@ from .keyboards import *
 from .middlewares import *
 from .routers import *
 from .version import version
+
+API_URL = Connector.API_URL
+API_FILE_URL = Connector.API_FILE_URL