+++ rocketgram/api/__init__.py
@@ -197,6 +197,3 @@ from .voice_chat_ended import VoiceChatEnded
 from .voice_chat_participants_invited import VoiceChatParticipantsInvited
 from .voice_chat_started import VoiceChatStarted
 from .webhook_info import WebhookInfo
-
-API_URL = "https://api.telegram.org/bot%s/"
-API_FILE_URL = "https://api.telegram.org/file/bot%s/%s"